Build 6.9.18/20200427

+++ plugins/fabrik_element/databasejoin/databasejoin.php
@@ -111,6 +111,11 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 	protected $orderBy = '';
 
 	/**
+	 * @var bool
+	 */
+	public $_rawFilter = false;
+
+	/**
 	 * Create the SQL select 'name AS alias' segment for list/form queries
 	 *
 	 * @param   array &$aFields   array of element names
@@ -732,8 +737,10 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 
 		foreach ($aDdObjs as &$o)
 		{
-			// For values like '1"'
-			// $$$ hugh - added second two params so we set double_encode false
+			// decode first, to decode all hex entities (like &#39;)
+			$o->text = html_entity_decode($o->text, ENT_QUOTES | ENT_XML1, 'UTF-8');
+
+			// Encode if necessary
 			if ($this->getDisplayType() != 'radio' && $this->getDisplayType() != 'checkbox')
 			{
 				$o->text = htmlspecialchars($o->text, ENT_NOQUOTES, 'UTF-8', false);
@@ -1055,8 +1062,16 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 	{
 		$params         = $this->getParams();
 		$element        = $this->getElement();
+		$filterWhere    = trim($params->get('database_join_filter_where_sql', ''));
 		$whereAccess    = $params->get('database_join_where_access', 26);
-		$where          = ($this->mustApplyWhere($whereAccess, $element->id) && $incWhere) ? $params->get('database_join_where_sql') : '';
+		$where          = '';
+
+		// $$$ hugh - 5/13/2019 - filter where now overrides form where, if present
+		if (FArrayHelper::getValue($opts, 'mode', '') !== 'filter' || empty($filterWhere))
+		{
+			$where = ($this->mustApplyWhere($whereAccess, $element->id) && $incWhere) ? $params->get('database_join_where_sql') : '';
+		}
+
 		$join           = $this->getJoin();
 		$thisTableAlias = is_null($thisTableAlias) ? $join->table_join_alias : $thisTableAlias;
 		$repeatCounter  = FArrayHelper::getValue($opts, 'repeatCounter', 0);
@@ -1089,7 +1104,6 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 		 * but first cut it's safer to put it here (don't ask).
 		 */
 
-		$filterWhere = trim($params->get('database_join_filter_where_sql', ''));
 		if (FArrayHelper::getValue($opts, 'mode', '') === 'filter' && !empty($filterWhere))
 		{
 			if (preg_match('/(ORDER\s+BY)(.*)/is', $filterWhere, $matches))
@@ -1117,7 +1131,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 		}
 
 		$where = $w->parseMessageForRepeats($where, $data, $this, $repeatCounter);
-		$where = $w->parseMessageForPlaceHolder($where, $data, false);
+		$where = $w->parseMessageForPlaceHolder($where, $data, false, false, null, false);
 
 		if (!$query)
 		{
@@ -1519,7 +1533,15 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 			{
 				for ($i = 0; $i < count($defaultLabels); $i++)
 				{
-					$url               = $popUrl . FArrayHelper::getValue($defaultValues, $i, '');
+					$value = FArrayHelper::getValue($defaultValues, $i, '');
+
+					// joins in repeat groups
+					if (is_object($value))
+					{
+						$value = FArrayHelper::getValue($value, 0, '');
+					}
+
+					$url               = $popUrl . $value;
 					$defaultLabels[$i] = '<a href="' . JRoute::_($url) . '">' . FArrayHelper::getValue($defaultLabels, $i) . '</a>';
 				}
 			}
@@ -1541,7 +1563,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 			return false;
 		}
 
-		$db    = $this->getDb();
+		$db = FabrikWorker::getDbo(true);
 		$query = $db->getQuery(true);
 		$query->select('id')->from('#__{package}_lists')->where('form_id =' . $popupFormId);
 		$db->setQuery($query);
@@ -2244,7 +2266,18 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 
 			foreach ($rows as &$r)
 			{
-				$r->text = strip_tags(FText::_($r->text));
+				// translate
+				$r->text = FText::_($r->text);
+
+				// decode first, to decode all hex entities (like &#39;)
+				$r->text = html_entity_decode($r->text, ENT_QUOTES | ENT_XML1, 'UTF-8');
+
+				// Encode if necessary
+				if (!in_array($element->get('filter_type'), array('checkbox')))
+				{
+					$r->text = strip_tags($r->text);
+					$r->text = htmlspecialchars($r->text, ENT_NOQUOTES, 'UTF-8', false);
+				}
 			}
 
 			$this->getFilterDisplayValues($default, $rows);
@@ -2372,6 +2405,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 
 			return $rows;
 		}
+
 		/*
 		 * list of all tables that have been joined to -
 		* if duplicated then we need to join using a table alias
@@ -2406,7 +2440,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 		}
 		else
 		{
-			if (!preg_match('/^WHERE\s+/i', $where))
+			if (!preg_match('/^\s*WHERE\s+/i', $where))
 			{
 				$where = 'WHERE ' . $where;
 			}
@@ -2683,11 +2717,12 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 		else
 		{
 			$group = $this->getGroup();
+			$exactMatch = $this->getElement()->get('filter_exact_match');
 
 			if (!$group->isJoin() && $group->canRepeat())
 			{
 				// Deprecated I think - repeat groups are always joins.
-				$fval = $this->getElement()->get('filter_exact_match') ? $originalValue : $value;
+				$fval = $exactMatch ? $originalValue : $value;
 				$str  = " ($key = $fval OR $key LIKE \"$originalValue',%\"" . " OR $key LIKE \"%:'$originalValue',%\""
 					. " OR $key LIKE \"%:'$originalValue'\"" . " )";
 			}
@@ -2707,11 +2742,40 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 						$where = $this->_db->qn($dbName . '.' . $this->getJoinValueFieldName());
 					}
 
-					$groupBy = $this->_db->qn($dbName . '.parent_id');
-					$rows    = $this->checkboxRows($groupBy, $condition, $value, $where);
-					$joinIds = array_keys($rows);
+					if (($fType === 'multiselect' || $fType === 'checkbox') && $exactMatch)
+					{
+						$rows    = $this->checkboxRows('id', $condition, $value, $where);
+						$joinIds = array();
+						$matches = array();
 
-					if (!empty($rows))
+						foreach ($rows as $row)
+						{
+							if (!array_key_exists($row->parent_id, $matches))
+							{
+								$matches[$row->parent_id] = array();
+							}
+
+							$matches[$row->parent_id][] = $row->value;
+						}
+
+						foreach ($matches as $joinId => $match)
+						{
+							$diff = array_diff($originalValue, $match);
+
+							if (empty($diff))
+							{
+								$joinIds[] = $joinId;
+							}
+						}
+					}
+					else
+					{
+						$groupBy = $this->_db->qn($dbName . '.parent_id');
+						$rows    = $this->checkboxRows($groupBy, $condition, $value, $where);
+						$joinIds = array_keys($rows);
+					}
+
+					if (!empty($joinIds))
 					{
 						// Either look for the ids in the main fabrik list or the group's list.
 						$groupJoinModel = $group->getJoinModel();
@@ -2804,9 +2868,9 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 
 		$db->setQuery($query, $offset, $limit);
 		$sql     = (string) $query;
+		FabrikHelperHTML::debug($sql, 'join checkboxRows:');
 		$groupBy = FabrikString::shortColName($groupBy);
 		$rows    = $db->loadObjectList($groupBy);
-		//ksort($rows);
 
 		return $rows;
 	}
@@ -3031,7 +3095,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 		$modalId                  = 'dbjoin_popupform';
 		$opts                     = $this->_getOptionVals();
 		$data                     = $this->getFormModel()->data;
-		$arSelected               = $this->getValue($data, $repeatCounter);
+		$arSelected               = $this->getValue($data, $repeatCounter, array('raw' => true));
 		$table                    = $params->get('join_db_name');
 		$opts                     = $this->getElementJSOptions($repeatCounter);
 		$forms                    = $this->getLinkedForms();
@@ -3095,6 +3159,14 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 				}
 
 				$m = str_replace(array('{', '}'), '', $m);
+
+				// remove any ||default
+                $bits = explode('||', $m);
+
+                if (count($bits) > 1)
+                {
+                    $m = $bits[0];
+                }
 
 				// $$$ hugh - we need to knock any _raw off, so JS can match actual element ID
 				$m = preg_replace('#_raw$#', '', $m);
@@ -3171,6 +3243,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 		$this->id = $this->app->input->getInt('element_id');
 		$this->loadMeForAjax();
 		$this->getElement(true);
+		$params = $this->getParams();
 		$filter  = JFilterInput::getInstance();
 		$request = $filter->clean($_REQUEST, 'array');
 		$groupModel = $this->getGroupModel();
@@ -3191,7 +3264,26 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 			}
 		}
 
-		echo json_encode($this->_getOptions($request));
+		$options = $this->_getOptions($request);
+		$eval = $params->get('databasejoin_where_ajax_default_eval');
+
+		if (!empty($eval))
+		{
+			$default = eval($eval);
+
+			if (!empty($default))
+			{
+				foreach ($options as $opt)
+				{
+					if ($opt->value === $default)
+					{
+						$opt->selected = true;
+					}
+				}
+			}
+		}
+
+		echo json_encode($options);
 	}
 
 	/**
@@ -3603,7 +3695,7 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 	private function _autocompleteWhere($how, $field, $search)
 	{
 		$db     = FabrikWorker::getDbo();
-		$search = strtolower($search);
+		$search = mb_strtolower($search);
 		$field  = 'LOWER(' . $field . ')';
 
 		switch ($how)
@@ -4053,9 +4145,21 @@ class PlgFabrik_ElementDatabasejoin extends PlgFabrik_ElementList
 			$key = $db->q($key);
 		});
 
-		$query = $db->getQuery(true);
-		$query->delete($db->qn($join->table_join))->where('id IN (' . implode(',', $keys) .')');
+		if (!empty($keys))
+		{
+			$query = $db->getQuery(true);
+			$query->delete($db->qn($join->table_join))->where('id IN (' . implode(',', $keys) . ')');
 
-		return $db->setQuery($query)->execute();
+			try
+			{
+				$db->setQuery($query)->execute();
+			}
+			catch (Exception $e)
+			{
+				// meh
+			}
+		}
+
+		return true;
 	}
 }
